+++ tests/testthat/test-markdown.R
@@ -619,3 +619,20 @@ test_that("markup in headings", {
     )
   )
 })
+
+test_that("alternative knitr engines", {
+  expect_snapshot(
+    print(out1 <- roc_proc_text(rd_roclet(), "
+      #' Title
+      #'
+      #' Description.
+      #'
+      #' ```{verbatim}
+      #' #| file = testthat::test_path(\"example.Rmd\")
+      #' ```
+      #' @md
+      #' @name x
+      NULL
+    "))
+  )
+})